// proteger_rutas/src/types/products.ts
import type { ICategory } from "./category";

export interface Product {
    id: number;
    eliminado: boolean;
    createdAt: string;
    nombre: string;
    precio: number;
    descripcion: string;
    stock: number;
    imagen: string;
    disponible: boolean;
    categorias: ICategory[];
}

export interface CartItem extends Product {
    cantidad: number;
}